Disabled destroy functions in ts
