+++ packages/appengine/src/loadBalancer/details/details.controller.ts
@@ -4,8 +4,7 @@ import { module } from 'angular';
 import type { IModalService } from 'angular-ui-bootstrap';
 import { cloneDeep } from 'lodash';
 
-import type { Application, ILoadBalancer, ILoadBalancerDeleteCommand } from '@spinnaker/core';
-import { ConfirmationModalService, LoadBalancerWriter } from '@spinnaker/core';
+import type { Application, ILoadBalancer } from '@spinnaker/core';
 import type { IAppengineLoadBalancer } from '../../domain';
 
 interface ILoadBalancerFromStateParams {
@@ -71,37 +70,6 @@ class AppengineLoadBalancerDetailsController implements IController {
           this.dispatchRules.push(rule.domain + rule.path);
         }
       });
-    }
-  }
-
-  private getConfirmationModalBodyHtml(): string {
-    const serverGroupNames = this.loadBalancer.serverGroups.map((serverGroup) => serverGroup.name);
-    const hasAny = serverGroupNames ? serverGroupNames.length > 0 : false;
-    const hasMoreThanOne = serverGroupNames ? serverGroupNames.length > 1 : false;
-
-    // HTML accepted by the confirmationModalService is static (i.e., not managed by angular).
-    if (hasAny) {
-      if (hasMoreThanOne) {
-        const listOfServerGroupNames = serverGroupNames.map((name) => `<li>${name}</li>`).join('');
-        return `<div class="alert alert-warning">
-            <p>
-              Deleting <b>${this.loadBalancer.name}</b> will destroy the following server groups:
-              <ul>
-                ${listOfServerGroupNames}
-              </ul>
-            </p>
-          </div>
-        `;
-      } else {
-        return `<div class="alert alert-warning">
-            <p>
-              Deleting <b>${this.loadBalancer.name}</b> will destroy <b>${serverGroupNames[0]}</b>.
-            </p>
-          </div>
-        `;
-      }
-    } else {
-      return null;
     }
   }
