+++ packages/appengine/src/serverGroup/details/details.html
@@ -35,37 +35,6 @@
           <li ng-if="ctrl.serverGroup.disabled">
             <a href ng-click="ctrl.enableServerGroup()"> Enable </a>
           </li>
-          <li ng-if="!ctrl.serverGroup.disabled && ctrl.canDisableServerGroup()">
-            <a href ng-click="ctrl.disableServerGroup()"> Disable </a>
-          </li>
-          <li
-            ng-if="!ctrl.serverGroup.disabled && !ctrl.canDisableServerGroup()"
-            uib-tooltip="You cannot disable a server group if it is the
-                           only server group receiving traffic from a load balancer."
-            class="disabled"
-          >
-            <a href> Disable </a>
-          </li>
-          <li ng-if="ctrl.canDestroyServerGroup()">
-            <a href ng-click="ctrl.destroyServerGroup()">Destroy</a>
-          </li>
-          <li
-            ng-if="!ctrl.canDestroyServerGroup()"
-            uib-tooltip="You cannot destroy a server group if it is the only server group
-                           receiving traffic from a load balancer. You may be able to delete
-                           this server group's load balancer."
-            class="disabled"
-          >
-            <a href>Destroy</a>
-          </li>
-          <li
-            uib-tooltip="It is not possible to clone an App Engine server group's full
-                           launch configuration. However, clicking this button will allow
-                           you to deploy into this server group's cluster."
-            class="disabled"
-          >
-            <a>Clone</a>
-          </li>
         </ul>
       </div>
     </div>
