+++ packages/google/src/loadBalancer/configure/http/httpLoadBalancer.write.service.js
@@ -29,5 +29,5 @@ angular
       });
     }
 
-    return upsertLoadBalancers;
+    return { upsertLoadBalancers };
   });
