--- packages/google/src/loadBalancer/details/deleteModal/deleteModal.html
@@ -1,31 +1,0 @@
-<div modal-page class="confirmation-modal">
-  <task-monitor monitor="ctrl.taskMonitor"></task-monitor>
-  <form role="form" ng-if="!ctrl.submitting">
-    <modal-close dismiss="$dismiss()"></modal-close>
-    <div class="modal-header">
-      <h4 class="modal-title">Really delete {{ctrl.loadBalancer.name}}?</h4>
-    </div>
-    <div class="modal-body confirmation-modal" ng-if="ctrl.hasHealthChecks()">
-      <div class="row">
-        <div class="col-md-offset-2 col-md-8" style="padding-left: 2px">
-          <div class="checkbox">
-            <label> <input ng-model="ctrl.params.deleteHealthChecks" type="checkbox" />Delete health checks? </label>
-            <help-field
-              expand="true"
-              content="If the health checks associated with this load balancer
-                                 are being used by other resources, they will not be deleted,
-                                 whether this option is selected or not."
-            ></help-field>
-          </div>
-        </div>
-      </div>
-    </div>
-    <gce-footer
-      action="ctrl.submit()"
-      cancel="ctrl.cancel()"
-      is-valid="ctrl.isValid()"
-      account="ctrl.loadBalancer.account"
-      verification="ctrl.verification"
-    ></gce-footer>
-  </form>
-</div>
