+++ packages/pluginsdk-peerdeps/package.json
@@ -46,71 +46,27 @@
     "utf-8-validate": "5.0.3"
   },
   "peerDependenciesMeta": {
-    "@rollup/plugin-commonjs": {
-      "dev": true
-    },
-    "@rollup/plugin-json": {
-      "dev": true
-    },
-    "@rollup/plugin-node-resolve": {
-      "dev": true
-    },
-    "@rollup/plugin-replace": {
-      "dev": true
-    },
-    "@rollup/plugin-typescript": {
-      "dev": true
-    },
-    "@rollup/plugin-url": {
-      "dev": true
-    },
-    "@spinnaker/eslint-plugin": {
-      "dev": true
-    },
-    "@types/react": {
-      "dev": true
-    },
-    "bufferutil": {
-      "dev": true
-    },
-    "npm-run-all": {
-      "dev": true
-    },
-    "postcss": {
-      "dev": true
-    },
-    "prettier": {
-      "dev": true
-    },
-    "pretty-quick": {
-      "dev": true
-    },
-    "rollup": {
-      "dev": true
-    },
-    "rollup-plugin-external-globals": {
-      "dev": true
-    },
-    "rollup-plugin-less": {
-      "dev": true
-    },
-    "rollup-plugin-postcss": {
-      "dev": true
-    },
-    "rollup-plugin-terser": {
-      "dev": true
-    },
-    "rollup-plugin-visualizer": {
-      "dev": true
-    },
-    "shx": {
-      "dev": true
-    },
-    "typescript": {
-      "dev": true
-    },
-    "utf-8-validate": {
-      "dev": true
-    }
+    "@rollup/plugin-commonjs": {},
+    "@rollup/plugin-json": {},
+    "@rollup/plugin-node-resolve": {},
+    "@rollup/plugin-replace": {},
+    "@rollup/plugin-typescript": {},
+    "@rollup/plugin-url": {},
+    "@spinnaker/eslint-plugin": {},
+    "@types/react": {},
+    "bufferutil": {},
+    "npm-run-all": {},
+    "postcss": {},
+    "prettier": {},
+    "pretty-quick": {},
+    "rollup": {},
+    "rollup-plugin-external-globals": {},
+    "rollup-plugin-less": {},
+    "rollup-plugin-postcss": {},
+    "rollup-plugin-terser": {},
+    "rollup-plugin-visualizer": {},
+    "shx": {},
+    "typescript": {},
+    "utf-8-validate": {}
   }
 }
